Settings menu now modifies settings file when the default path is changed. The default path is also loaded from the file at startup.
The add menu is now functional. A folder is created in the default path named after the note title. An info file is created containing tags, description, title, and any other information we'll need to store in the future. The files that were selected using the JFileChooser are copied into the folder. They are titled n.ext, where n is the order in which the file was chosen in the JFileChooser. The original extension is kept.